RSWW-72 payment-service: send dtos converted to json

+++ payment-service/src/main/java/org/microarchitecturovisco/paymentservice/services/PaymentService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 public class PaymentService {
 
     public HandlePaymentResponseDto verifyTransaction(HandlePaymentRequestDto requestDto) {
-        boolean transactionApproved = Integer.parseInt(requestDto.getCardNumber()) % 2 == 0;
+        boolean transactionApproved = Long.parseLong(requestDto.getCardNumber()) % 2 == 0;
 
         return HandlePaymentResponseDto.builder()
                 .id(requestDto.getId())
